flask-retful-api/requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
requests==2.22.0
six==1.14.0
urllib3==1.25.8
Werkzeug==0.16.1
pandas==1.0.5
ujson==3.0.0
flask_migrate==2.5.3
xlrd==1.2.0
gevent==20.6.2
Flask-Script==2.0.6
gunicorn==20.0.4
gevent==20.6.2
redis==3.4.1
numpy==1.19.1
opencv-python==4.2.0.32
Pillow==7.2.0
flask-restx==0.2.0
tencentcloud-sdk-python==3.0.250
zxing==0.11
gmssl==3.2.1
pyzbar==0.1.8
pdf2image==1.12.1






